@@ -11,7 +11,8 @@
   </modules>
   
   <properties>
-    <maven.compiler.release>16</maven.compiler.release>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
     <lwjgl.version>3.2.3</lwjgl.version>
     <joml.version>1.10.1</joml.version>
     <lwjgl.natives>natives-windows</lwjgl.natives>
@@ -22,7 +23,7 @@
     <plugins>
       <plugin>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.8.0</version>
+        <version>3.8.1</version>
       </plugin>
     </plugins>
   </build>
