@@ -1,6 +1,6 @@
 package com.avogine.ecs;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
 import java.io.*;
 
